[YS-228] refactor: import 경로 수정
